feat: Integrate Hilt and Firebase
This commit integrates Hilt for dependency injection and
Firebase for database interaction.  It adds the AppModule
for providing dependencies like Context, Gson, and
FirebaseDatabase.  It also creates the RingoPhoneApplication
class and configures MainActivity to write "Hello Abdo!" to the
Firebase database for testing the connection. Permissions
are added to the manifest. The MainActivity file is also moved
to a new package.